# Optimized accuracy calculation