@shaibagon 2017-08-09
I think what can really help here is a GPU implementation for this layer...