@shaibagon 2017-08-10
Look at PR#5836. 
In order to estimate `top_k` accuracy one does not need to sort / partially sort / priority_queue the scores AT ALL! 
We know **in advance** what is the predicted score for the "true label", we only need to go once over all the other predictions and make sure there are no more than `top_k` better predictions.
This way, we can estimate accuracy with O(L) (number of labels) per pixels without any fancy data structures.