@Noiredd 2017-08-21
@shaibagon Indeed, we don't need to know _what did the network answer_, only whether the correct label was among the predictions - and without sorts/queues it's so much easier to do on GPU. This is pretty clever ;)
Closing this in favor of #5386